refactor(provider): changed 'implementation' pkg and struct for 'provider' and updated code in other pkg's
